refactor: card 객체 생성
