+++ src/main/kotlin/blackjack/domain/card/Suit.kt
@@ -1,8 +1,8 @@
-package blackjack.domain
+package blackjack.domain.card
 
 enum class Suit(val symbol: String) {
     HEARTS("하트"),
     DIAMONDS("다이아몬드"),
-    CLOVERS("클로버"),
+    CLUBS("클로버"),
     SPADES("스페이드"),
 }
